:truck: move only the platform-specific part of log tailing into platform-specific files
