+++ pkg/app/app.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
+	"github.com/aybabtme/humanlog"
 	"github.com/jesseduffield/lazygit/pkg/commands"
 	"github.com/jesseduffield/lazygit/pkg/commands/oscommands"
 	"github.com/jesseduffield/lazygit/pkg/config"
@@ -305,4 +306,26 @@ func (app *App) KnownError(err error) (string, bool) {
 		}
 	}
 	return "", false
+}
+
+func TailLogs() {
+	logFilePath, err := config.LogPath()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Printf("Tailing log file %s\n\n", logFilePath)
+
+	opts := humanlog.DefaultOptions
+	opts.Truncates = false
+
+	_, err = os.Stat(logFilePath)
+	if err != nil {
+		if os.IsNotExist(err) {
+			log.Fatal("Log file does not exist. Run `lazygit --debug` first to create the log file")
+		}
+		log.Fatal(err)
+	}
+
+	TailLogsForPlatform(logFilePath, opts)
 }
